+++ src/vite/index.ts
@@ -1,0 +1,1 @@
+export { spriteSheetBuilder } from "./spriteSheetBuilderPlugin.ts";